moved fix to lib.rs, tests are still wip
